requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
django-pagedown==2.2.1
django-password-reset==2.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
mailchimp-transactional==1.0.50
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13